+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
+## 0.0.4
 
-* TODO: Describe initial release.
+* Resolved cursor position bugs.
+* Resolved bug of numbers position in urdu and sindhi text field.
+* Updated numeric key text when keyboard is changing.